requirements.txt
amqp==2.5.1
appnope==0.1.0
Babel==2.7.0
backcall==0.1.0
backports.csv==1.0.7
beautifulsoup4==4.8.1
billiard==3.6.1.0
bs4==0.0.1
celery==4.3.0
certifi==2019.9.11
chardet==3.0.4
decorator==4.4.1
defusedxml==0.6.0
diff-match-patch==20181111
Django==2.2.3
django-celery-beat==1.5.0
django-celery-results==1.1.2
django-crispy-forms==1.7.2
django-extensions==2.2.5
django-formtools==2.1
django-import-export==1.2.0
django-reversion==3.0.4
django-timezone-field==3.0
django-xadmin==0.5.0
djangorestframework==3.9.4
djangorestframework-jwt==1.11.0
et-xmlfile==1.0.1
flower==0.9.3
future==0.17.1
httplib2==0.9.2
idna==2.8
importlib-metadata==0.19
ipython==7.9.0
ipython-genutils==0.2.0
jdcal==1.4.1
jedi==0.15.1
kombu==4.6.4
lxml==4.4.1
mysqlclient==1.4.2.post1
numpy==1.16.4
odfpy==1.4.0
openpyxl==2.6.2
pandas==0.25.1
parso==0.5.1
pexpect==4.7.0
pickleshare==0.7.5
prompt-toolkit==2.0.10
ptyprocess==0.6.0
Pygments==2.4.2
PyJWT==1.7.1
PyMySQL==0.9.3
python-crontab==2.3.9
python-dateutil==2.8.0
pytz==2019.1
PyYAML==5.1.1
requests==2.22.0
six==1.12.0
soupsieve==1.9.4
SQLAlchemy==1.3.9
sqlparse==0.3.0
tablib==0.13.0
tornado==5.1.1
traitlets==4.3.3
tushare==1.2.48
urllib3==1.25.6
vine==1.3.0
wcwidth==0.1.7
xadmin==2.0.1
xlrd==1.2.0
xlwt==1.3.0
zipp==0.5.2
